Fix const version of first_unused_segment()
The const version of first_unused_segment() differs from the non-const
version and has identical return values for the ternary, which does
not make sense. The implementation is also wrong, since it always
gives a segment that will extend to the end of the container, even if
there is already data present there. The non-const implementation is
correct.